feat(useElastic): Add hook useElastic
